+++ src/main/java/car/racing/CarRacingFactory.java
@@ -1,5 +1,8 @@
 package car.racing;
 
+import car.racing.domain.Car;
+import car.racing.view.ResultView;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
