+++ src/main/java/car/racing/CarRacingMain.java
@@ -1,5 +1,7 @@
 package car.racing;
 
+import car.racing.view.InputView;
+
 public class CarRacingMain {
 
     public static void main(String[] args) {
