+++ src/test/java/car/racing/CarRacingManagerTest.java
@@ -1,5 +1,6 @@
 package car.racing;
 
+import car.racing.domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
