+++ src/test/java/car/racing/FakeResultView.java
@@ -1,5 +1,8 @@
 package car.racing;
 
+import car.racing.domain.Car;
+import car.racing.view.ResultViewContract;
+
 import java.util.ArrayList;
 import java.util.List;
